requirements.txt
torch>=2.0
torchvision>=0.15
open_clip_torch>=2.24
pillow>=10.0
numpy>=1.24,<2
scikit-learn>=1.2
tqdm>=4.66
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
jinja2>=3.1
gdown